requirements.txt
# Core Django
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5

# Database
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0
dj-database-url==2.1.0

# Firebase
firebase-admin==6.2.0
google-cloud-storage==2.10.0
PyJWT==2.8.0
cryptography==41.0.7

# File Processing
pillow==10.1.0
python-magic==0.4.27

# API Documentation
drf-spectacular==0.27.0

# Environment
python-decouple==3.8
whitenoise==6.6.0

# Development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.5.0

# Production
gunicorn==21.2.0
